@@ -5,7 +5,6 @@ import (
 	"context"
 	"fmt"
 	"io/fs"
-	"io/ioutil"
 	"net/http"
 	"os"
 	"strings"
@@ -38,17 +37,17 @@ import (
 func TestCreateLocalDir(t *testing.T) {
 	dataPath := t.TempDir()
 	localFile := dataPath + "/local"
-	require.NoError(t, ioutil.WriteFile(localFile, []byte("d"), 0o644))
+	require.NoError(t, os.WriteFile(localFile, []byte("d"), 0o644))
 	_, err := New(context.Background(), Config{
 		DataPath:         dataPath,
 		MaxBlockDuration: 30 * time.Minute,
-	})
+	}, NoLimit)
 	require.Error(t, err)
 	require.NoError(t, os.Remove(localFile))
 	_, err = New(context.Background(), Config{
 		DataPath:         dataPath,
 		MaxBlockDuration: 30 * time.Minute,
-	})
+	}, NoLimit)
 	require.NoError(t, err)
 }
 
@@ -143,7 +142,7 @@ func TestMergeProfilesStacktraces(t *testing.T) {
 	db, err := New(context.Background(), Config{
 		DataPath:         testDir,
 		MaxBlockDuration: time.Duration(100000) * time.Minute, // we will manually flush
-	})
+	}, NoLimit)
 	require.NoError(t, err)
 	defer require.NoError(t, db.Close())
 
@@ -271,7 +270,7 @@ func TestMergeProfilesPprof(t *testing.T) {
 	db, err := New(context.Background(), Config{
 		DataPath:         testDir,
 		MaxBlockDuration: time.Duration(100000) * time.Minute, // we will manually flush
-	})
+	}, NoLimit)
 	require.NoError(t, err)
 	defer require.NoError(t, db.Close())
 
@@ -451,14 +450,8 @@ func TestFilterProfiles(t *testing.T) {
 	}, filtered)
 }
 
-type fakeBlock struct {
-	id   string
-	size uint64 // in mbytes
-}
-
 type fakeVolumeFS struct {
 	mock.Mock
-	blocks []fakeBlock
 }
 
 func (f *fakeVolumeFS) HasHighDiskUtilization(path string) (*diskutil.VolumeStats, error) {
@@ -470,6 +463,7 @@ func (f *fakeVolumeFS) Open(path string) (fs.File, error) {
 	args := f.Called(path)
 	return args[0].(fs.File), args.Error(1)
 }
+
 func (f *fakeVolumeFS) RemoveAll(path string) error {
 	args := f.Called(path)
 	return args.Error(0)
