@@ -16,6 +16,7 @@ import (
 	"github.com/bufbuild/connect-go"
 	"github.com/dustin/go-humanize"
 	"github.com/go-kit/log"
+	"github.com/grafana/dskit/kv"
 	"github.com/grafana/dskit/ring"
 	"github.com/grafana/dskit/ring/client"
 	"github.com/grafana/dskit/services"
@@ -28,16 +29,27 @@ import (
 	"github.com/grafana/phlare/pkg/ingester/clientpool"
 	"github.com/grafana/phlare/pkg/tenant"
 	"github.com/grafana/phlare/pkg/testhelper"
+	"github.com/grafana/phlare/pkg/validation"
 )
+
+var ringConfig = RingConfig{
+	KVStore:      kv.Config{Store: "inmemory"},
+	InstanceID:   "foo",
+	InstancePort: int(8080),
+	InstanceAddr: "127.0.0.1",
+	ListenPort:   int(8080),
+}
 
 func Test_ConnectPush(t *testing.T) {
 	mux := http.NewServeMux()
 	ing := newFakeIngester(t, false)
-	d, err := New(Config{}, testhelper.NewMockRing([]ring.InstanceDesc{
+	d, err := New(Config{
+		DistributorRing: ringConfig,
+	}, testhelper.NewMockRing([]ring.InstanceDesc{
 		{Addr: "foo"},
 	}, 3), func(addr string) (client.PoolClient, error) {
 		return ing, nil
-	}, nil, log.NewLogfmtLogger(os.Stdout))
+	}, newOverrides(t), nil, log.NewLogfmtLogger(os.Stdout))
 
 	require.NoError(t, err)
 	mux.Handle(pushv1connect.NewPusherServiceHandler(d, connect.WithInterceptors(tenant.NewAuthInterceptor(true))))
@@ -51,6 +63,7 @@ func Test_ConnectPush(t *testing.T) {
 			{
 				Labels: []*typesv1.LabelPair{
 					{Name: "cluster", Value: "us-central1"},
+					{Name: "__name__", Value: "cpu"},
 				},
 				Samples: []*pushv1.RawSample{
 					{
@@ -77,6 +90,7 @@ func Test_Replication(t *testing.T) {
 			{
 				Labels: []*typesv1.LabelPair{
 					{Name: "cluster", Value: "us-central1"},
+					{Name: "__name__", Value: "cpu"},
 				},
 				Samples: []*pushv1.RawSample{
 					{
@@ -86,13 +100,13 @@ func Test_Replication(t *testing.T) {
 			},
 		},
 	})
-	d, err := New(Config{}, testhelper.NewMockRing([]ring.InstanceDesc{
+	d, err := New(Config{DistributorRing: ringConfig}, testhelper.NewMockRing([]ring.InstanceDesc{
 		{Addr: "1"},
 		{Addr: "2"},
 		{Addr: "3"},
 	}, 3), func(addr string) (client.PoolClient, error) {
 		return ingesters[addr], nil
-	}, nil, log.NewLogfmtLogger(os.Stdout))
+	}, newOverrides(t), nil, log.NewLogfmtLogger(os.Stdout))
 	require.NoError(t, err)
 	// only 1 ingester failing should be fine.
 	resp, err := d.Push(ctx, req)
@@ -108,12 +122,13 @@ func Test_Replication(t *testing.T) {
 func Test_Subservices(t *testing.T) {
 	ing := newFakeIngester(t, false)
 	d, err := New(Config{
-		PoolConfig: clientpool.PoolConfig{ClientCleanupPeriod: 1 * time.Second},
+		PoolConfig:      clientpool.PoolConfig{ClientCleanupPeriod: 1 * time.Second},
+		DistributorRing: ringConfig,
 	}, testhelper.NewMockRing([]ring.InstanceDesc{
 		{Addr: "foo"},
 	}, 1), func(addr string) (client.PoolClient, error) {
 		return ing, nil
-	}, nil, log.NewLogfmtLogger(os.Stdout))
+	}, newOverrides(t), nil, log.NewLogfmtLogger(os.Stdout))
 
 	require.NoError(t, err)
 	require.NoError(t, d.StartAsync(context.Background()))
@@ -164,4 +179,107 @@ func TestBuckets(t *testing.T) {
 	for _, r := range prometheus.ExponentialBucketsRange(10*1024, 15*1024*1024, 30) {
 		t.Log(humanize.Bytes(uint64(r)))
 	}
+}
+
+func Test_Limits(t *testing.T) {
+	mux := http.NewServeMux()
+	ing := newFakeIngester(t, false)
+	d, err := New(Config{
+		DistributorRing: ringConfig,
+	}, testhelper.NewMockRing([]ring.InstanceDesc{
+		{Addr: "foo"},
+	}, 3), func(addr string) (client.PoolClient, error) {
+		return ing, nil
+	}, newOverrides(t), nil, log.NewLogfmtLogger(os.Stdout))
+
+	require.NoError(t, err)
+	mux.Handle(pushv1connect.NewPusherServiceHandler(d, connect.WithInterceptors(tenant.NewAuthInterceptor(true))))
+	s := httptest.NewServer(mux)
+	defer s.Close()
+
+	client := pushv1connect.NewPusherServiceClient(http.DefaultClient, s.URL, connect.WithInterceptors(tenant.NewAuthInterceptor(true)))
+
+	t.Run("rate_limit", func(t *testing.T) {
+		resp, err := client.Push(tenant.InjectTenantID(context.Background(), "user-1"), connect.NewRequest(&pushv1.PushRequest{
+			Series: []*pushv1.RawProfileSeries{
+				{
+					Labels: []*typesv1.LabelPair{
+						{Name: "cluster", Value: "us-central1"},
+						{Name: "__name__", Value: "cpu"},
+					},
+					Samples: []*pushv1.RawSample{
+						{
+							RawProfile: testProfile(t),
+						},
+					},
+				},
+			},
+		}))
+		require.Error(t, err)
+		require.Equal(t, connect.CodeResourceExhausted, connect.CodeOf(err))
+		require.Nil(t, resp)
+	})
+	t.Run("label limit", func(t *testing.T) {
+		resp, err := client.Push(tenant.InjectTenantID(context.Background(), "user-1"), connect.NewRequest(&pushv1.PushRequest{
+			Series: []*pushv1.RawProfileSeries{
+				{
+					Labels: []*typesv1.LabelPair{
+						{Name: "clusterdddwqdqdqdqdqdqw", Value: "us-central1"},
+						{Name: "__name__", Value: "cpu"},
+					},
+					Samples: []*pushv1.RawSample{
+						{
+							RawProfile: testProfile(t),
+						},
+					},
+				},
+			},
+		}))
+		require.Error(t, err)
+		require.Equal(t, connect.CodeInvalidArgument, connect.CodeOf(err))
+		require.Nil(t, resp)
+	})
+}
+
+func TestBadPushRequest(t *testing.T) {
+	mux := http.NewServeMux()
+	ing := newFakeIngester(t, false)
+	d, err := New(Config{
+		DistributorRing: ringConfig,
+	}, testhelper.NewMockRing([]ring.InstanceDesc{
+		{Addr: "foo"},
+	}, 3), func(addr string) (client.PoolClient, error) {
+		return ing, nil
+	}, newOverrides(t), nil, log.NewLogfmtLogger(os.Stdout))
+
+	require.NoError(t, err)
+	mux.Handle(pushv1connect.NewPusherServiceHandler(d, connect.WithInterceptors(tenant.NewAuthInterceptor(true))))
+	s := httptest.NewServer(mux)
+	defer s.Close()
+
+	client := pushv1connect.NewPusherServiceClient(http.DefaultClient, s.URL, connect.WithInterceptors(tenant.NewAuthInterceptor(true)))
+
+	_, err = client.Push(tenant.InjectTenantID(context.Background(), "foo"), connect.NewRequest(&pushv1.PushRequest{
+		Series: []*pushv1.RawProfileSeries{
+			{
+				Labels: []*typesv1.LabelPair{
+					{Name: "cluster", Value: "us-central1"},
+					{Name: "__name__", Value: "cpu"},
+				},
+			},
+		},
+	}))
+	require.Error(t, err)
+	require.Equal(t, connect.CodeInvalidArgument, connect.CodeOf(err))
+}
+
+func newOverrides(t *testing.T) *validation.Overrides {
+	t.Helper()
+	return validation.MockOverrides(func(defaults *validation.Limits, tenantLimits map[string]*validation.Limits) {
+		l := validation.MockDefaultLimits()
+		l.IngestionRateMB = 0.0150
+		l.IngestionBurstSizeMB = 0.0015
+		l.MaxLabelNameLength = 10
+		tenantLimits["user-1"] = l
+	})
 }
