@@ -12,6 +12,7 @@ import (
 	"github.com/grafana/dskit/kv/codec"
 	"github.com/grafana/dskit/kv/memberlist"
 	"github.com/grafana/dskit/ring"
+	"github.com/grafana/dskit/runtimeconfig"
 	"github.com/grafana/dskit/services"
 	grpcgw "github.com/grpc-ecosystem/grpc-gateway/v2/runtime"
 	"github.com/pkg/errors"
@@ -28,14 +29,14 @@ import (
 
 	agentv1 "github.com/grafana/phlare/api/gen/proto/go/agent/v1"
 	"github.com/grafana/phlare/api/gen/proto/go/agent/v1/agentv1connect"
-	ingesterv1connect "github.com/grafana/phlare/api/gen/proto/go/ingester/v1/ingesterv1connect"
+	"github.com/grafana/phlare/api/gen/proto/go/ingester/v1/ingesterv1connect"
 	"github.com/grafana/phlare/api/gen/proto/go/push/v1/pushv1connect"
 	"github.com/grafana/phlare/api/gen/proto/go/querier/v1/querierv1connect"
 	statusv1 "github.com/grafana/phlare/api/gen/proto/go/status/v1"
 	"github.com/grafana/phlare/api/openapiv2"
 	"github.com/grafana/phlare/pkg/agent"
 	"github.com/grafana/phlare/pkg/distributor"
-	frontend "github.com/grafana/phlare/pkg/frontend"
+	"github.com/grafana/phlare/pkg/frontend"
 	"github.com/grafana/phlare/pkg/frontend/frontendpb/frontendpbconnect"
 	"github.com/grafana/phlare/pkg/ingester"
 	objstoreclient "github.com/grafana/phlare/pkg/objstore/client"
@@ -48,32 +49,30 @@ import (
 	"github.com/grafana/phlare/pkg/usagestats"
 	"github.com/grafana/phlare/pkg/util"
 	"github.com/grafana/phlare/pkg/util/build"
+	"github.com/grafana/phlare/pkg/validation"
+	"github.com/grafana/phlare/pkg/validation/exporter"
 )
 
 // The various modules that make up Phlare.
 const (
-	All            string = "all"
-	Agent          string = "agent"
-	Distributor    string = "distributor"
-	Server         string = "server"
-	Ring           string = "ring"
-	Ingester       string = "ingester"
-	MemberlistKV   string = "memberlist-kv"
-	Querier        string = "querier"
-	GRPCGateway    string = "grpc-gateway"
-	Storage        string = "storage"
-	UsageReport    string = "usage-stats"
-	QueryFrontend  string = "query-frontend"
-	QueryScheduler string = "query-scheduler"
+	All               string = "all"
+	Agent             string = "agent"
+	Distributor       string = "distributor"
+	Server            string = "server"
+	Ring              string = "ring"
+	Ingester          string = "ingester"
+	MemberlistKV      string = "memberlist-kv"
+	Querier           string = "querier"
+	GRPCGateway       string = "grpc-gateway"
+	Storage           string = "storage"
+	UsageReport       string = "usage-stats"
+	QueryFrontend     string = "query-frontend"
+	QueryScheduler    string = "query-scheduler"
+	RuntimeConfig     string = "runtime-config"
+	Overrides         string = "overrides"
+	OverridesExporter string = "overrides-exporter"
 
-	// RuntimeConfig            string = "runtime-config"
-	// Overrides                string = "overrides"
-	// OverridesExporter        string = "overrides-exporter"
-	// TenantConfigs            string = "tenant-configs"
 	// QueryFrontendTripperware string = "query-frontend-tripperware"
-	// RulerStorage             string = "ruler-storage"
-	// Ruler                    string = "ruler"
-	// TableManager             string = "table-manager"
 	// Compactor                string = "compactor"
 	// IndexGateway             string = "index-gateway"
 	// IndexGatewayRing         string = "index-gateway-ring"
@@ -104,14 +103,63 @@ func (f *Phlare) initQueryFrontend() (services.Service, error) {
 	return frontendSvc, nil
 }
 
-type fakeLimits struct{}
+func (f *Phlare) initRuntimeConfig() (services.Service, error) {
+	if len(f.Cfg.RuntimeConfig.LoadPath) == 0 {
+		// no need to initialize module if load path is empty
+		return nil, nil
+	}
 
-func (fakeLimits) MaxQueriersPerUser(user string) int { return 0 }
+	f.Cfg.RuntimeConfig.Loader = loadRuntimeConfig
+
+	// make sure to set default limits before we start loading configuration into memory
+	validation.SetDefaultLimitsForYAMLUnmarshalling(f.Cfg.LimitsConfig)
+
+	serv, err := runtimeconfig.New(f.Cfg.RuntimeConfig, prometheus.WrapRegistererWithPrefix("phlare_", f.reg), log.With(f.logger, "component", "runtime-config"))
+	if err == nil {
+		// TenantLimits just delegates to RuntimeConfig and doesn't have any state or need to do
+		// anything in the start/stopping phase. Thus we can create it as part of runtime config
+		// setup without any service instance of its own.
+		f.TenantLimits = newTenantLimits(serv)
+	}
+
+	f.RuntimeConfig = serv
+
+	f.Server.HTTP.Methods("GET").Path("/runtime_config").Handler(runtimeConfigHandler(f.RuntimeConfig, f.Cfg.LimitsConfig))
+	f.Server.HTTP.Methods("GET").Path("/api/v1/tenant_limits").Handler(middleware.AuthenticateUser.Wrap(validation.TenantLimitsHandler(f.Cfg.LimitsConfig, f.TenantLimits)))
+	return serv, err
+}
+
+func (f *Phlare) initOverrides() (serv services.Service, err error) {
+	f.Overrides, err = validation.NewOverrides(f.Cfg.LimitsConfig, f.TenantLimits)
+	// overrides don't have operational state, nor do they need to do anything more in starting/stopping phase,
+	// so there is no need to return any service.
+	return nil, err
+}
+
+func (f *Phlare) initOverridesExporter() (services.Service, error) {
+	overridesExporter, err := exporter.NewOverridesExporter(
+		f.Cfg.OverridesExporter,
+		&f.Cfg.LimitsConfig,
+		f.TenantLimits,
+		log.With(f.logger, "component", "overrides-exporter"),
+		f.reg,
+	)
+	if err != nil {
+		return nil, errors.Wrap(err, "failed to instantiate overrides-exporter")
+	}
+	if f.reg != nil {
+		f.reg.MustRegister(overridesExporter)
+	}
+
+	f.Server.HTTP.Methods("GET", "POST").Path("/overrides-exporter/ring").HandlerFunc(overridesExporter.RingHandler)
+
+	return overridesExporter, nil
+}
 
 func (f *Phlare) initQueryScheduler() (services.Service, error) {
 	f.Cfg.QueryScheduler.ServiceDiscovery.SchedulerRing.ListenPort = f.Cfg.Server.HTTPListenPort
 
-	s, err := scheduler.NewScheduler(f.Cfg.QueryScheduler, &fakeLimits{}, log.With(f.logger, "component", "scheduler"), f.reg)
+	s, err := scheduler.NewScheduler(f.Cfg.QueryScheduler, f.Overrides, log.With(f.logger, "component", "scheduler"), f.reg)
 	if err != nil {
 		return nil, errors.Wrap(err, "query-scheduler init")
 	}
@@ -195,7 +243,8 @@ func (f *Phlare) initGRPCGateway() (services.Service, error) {
 }
 
 func (f *Phlare) initDistributor() (services.Service, error) {
-	d, err := distributor.New(f.Cfg.Distributor, f.ring, nil, f.reg, log.With(f.logger, "component", "distributor"), f.auth)
+	f.Cfg.Distributor.DistributorRing.ListenPort = f.Cfg.Server.HTTPListenPort
+	d, err := distributor.New(f.Cfg.Distributor, f.ring, nil, f.Overrides, f.reg, log.With(f.logger, "component", "distributor"), f.auth)
 	if err != nil {
 		return nil, err
 	}
@@ -204,6 +253,8 @@ func (f *Phlare) initDistributor() (services.Service, error) {
 	f.pusherClient = d
 
 	pushv1connect.RegisterPusherServiceHandler(f.Server.HTTP, d, f.auth)
+	f.Server.HTTP.Path("/distributor/ring").Methods("GET", "POST").Handler(d)
+
 	return d, nil
 }
 
@@ -241,8 +292,10 @@ func (f *Phlare) initMemberlistKV() (services.Service, error) {
 
 	f.MemberlistKV = memberlist.NewKVInitService(&f.Cfg.MemberlistKV, f.logger, dnsProvider, f.reg)
 
+	f.Cfg.Distributor.DistributorRing.KVStore.MemberlistKV = f.MemberlistKV.GetMemberlistKV
 	f.Cfg.Ingester.LifecyclerConfig.RingConfig.KVStore.MemberlistKV = f.MemberlistKV.GetMemberlistKV
 	f.Cfg.QueryScheduler.ServiceDiscovery.SchedulerRing.KVStore.MemberlistKV = f.MemberlistKV.GetMemberlistKV
+	f.Cfg.OverridesExporter.Ring.KVStore.MemberlistKV = f.MemberlistKV.GetMemberlistKV
 
 	f.Cfg.Frontend.QuerySchedulerDiscovery = f.Cfg.QueryScheduler.ServiceDiscovery
 	f.Cfg.Worker.QuerySchedulerDiscovery = f.Cfg.QueryScheduler.ServiceDiscovery
@@ -289,7 +342,7 @@ func (f *Phlare) context() context.Context {
 func (f *Phlare) initIngester() (_ services.Service, err error) {
 	f.Cfg.Ingester.LifecyclerConfig.ListenPort = f.Cfg.Server.HTTPListenPort
 
-	ingester, err := ingester.New(f.context(), f.Cfg.Ingester, f.Cfg.PhlareDB, f.storageBucket)
+	ingester, err := ingester.New(f.context(), f.Cfg.Ingester, f.Cfg.PhlareDB, f.storageBucket, f.Overrides)
 	if err != nil {
 		return nil, err
 	}
@@ -410,7 +463,6 @@ func (f *Phlare) initUsageReport() (services.Service, error) {
 
 type statusService struct {
 	statusv1.UnimplementedStatusServiceServer
-	configYaml    string
 	defaultConfig *Config
 	actualConfig  *Config
 }
