@@ -516,7 +516,7 @@ type limits struct {
 	queriers int
 }
 
-func (l limits) MaxQueriersPerUser(_ string) int {
+func (l limits) MaxQueriersPerTenant(_ string) int {
 	return l.queriers
 }
 
